+++ integrations/jira/integration.py
@@ -9,31 +9,14 @@ from port_ocean.core.handlers.port_app_config.models import (
 from port_ocean.core.integrations.base import BaseIntegration
 from pydantic.fields import Field
 
-SprintState = Literal["active", "closed", "future"]
-
 
 class JiraIssueSelector(Selector):
     jql: str | None = Field(
         description="Jira Query Language (JQL) query to filter issues",
     )
-    source: Literal["board", "sprint", "all"] = Field(
-        default="sprint",
-        description="Where issues are sourced from",
-    )
-    # when resyncing issues, there is no way to retrieve the config
-    # set for the `sprint` kind, so we need to duplicate the state
-    # field. This is redundant, but necessary.
-    state: SprintState = Field(
-        alias="sprintState",
-        default="active",
-        description="State of the sprint",
-    )
-
-
-class JiraSprintSelector(Selector):
-    state: SprintState = Field(
-        default="active",
-        description="State of the sprint",
+    fields: str | None = Field(
+        description="Additional fields to be included in the API response",
+        default="*all",
     )
 
 
@@ -42,15 +25,10 @@ class JiraIssueResourceConfig(ResourceConfig):
     selector: JiraIssueSelector
 
 
-class JiraSprintResourceConfig(ResourceConfig):
-    kind: Literal["sprint"]
-    selector: JiraSprintSelector
-
-
 class JiraPortAppConfig(PortAppConfig):
-    resources: list[
-        JiraIssueResourceConfig | JiraSprintResourceConfig | ResourceConfig
-    ] = Field(default_factory=list)
+    resources: list[JiraIssueResourceConfig | ResourceConfig] = Field(
+        default_factory=list
+    )
 
 
 class JiraIntegration(BaseIntegration):